Struct cargo_toml::Badges

pub struct Badges {
    pub appveyor: Option<Badge>,
    pub circle_ci: Option<Badge>,
    pub gitlab: Option<Badge>,
    pub travis_ci: Option<Badge>,
    pub codecov: Option<Badge>,
    pub coveralls: Option<Badge>,
    pub is_it_maintained_issue_resolution: Option<Badge>,
    pub is_it_maintained_open_issues: Option<Badge>,
    pub maintenance: Maintenance,
}

Deprecated

Fields
appveyor: Option<Badge>

Appveyor: repository is required. branch is optional; default is master service is optional; valid values are github (default), bitbucket, and gitlab; id is optional; you can specify the appveyor project id if you want to use that instead. project_name is optional; use when the repository name differs from the appveyor project name.

circle_ci: Option<Badge>

Circle CI: repository is required. branch is optional; default is master

gitlab: Option<Badge>

GitLab: repository is required. branch is optional; default is master

travis_ci: Option<Badge>
👎Deprecated: badges are deprecated, and travis is dead

Travis CI: repository in format "<user>/<project>" is required. branch is optional; default is master

codecov: Option<Badge>

Codecov: repository is required. branch is optional; default is master service is optional; valid values are github (default), bitbucket, and gitlab.

coveralls: Option<Badge>

Coveralls: repository is required. branch is optional; default is master service is optional; valid values are github (default) and bitbucket.

is_it_maintained_issue_resolution: Option<Badge>

Is it maintained resolution time: repository is required.

is_it_maintained_open_issues: Option<Badge>

Is it maintained percentage of open issues: repository is required.

maintenance: Maintenance

Maintenance: status is required. Available options are actively-developed, passively-maintained, as-is, experimental, looking-for-maintainer, deprecated, and the default none, which displays no badge on crates.io.