Trait cargo_toml::AbstractFilesystem

pub trait AbstractFilesystem {
    fn file_names_in(&self, rel_path: &str) -> Result<HashSet<Box<str>>>;

    fn read_root_workspace(
        &self,
        _rel_path_hint: Option<&str>
    ) -> Result<(Vec<u8>, PathBuf)> { ... }
}

Required Methods

fn file_names_in(&self, rel_path: &str) -> Result<HashSet<Box<str>>>

List all files and directories at the given relative path (no leading /).

Provided Methods

fn read_root_workspace(
&self,
_rel_path_hint: Option<&str>
) -> Result<(Vec<u8>, PathBuf)>

The rel_path_hint may be specified explicitly by package.workspace (it may be relative like "../", without Cargo.toml) or None, which means you have to search for workspace's Cargo.toml` in parent directories.

Read bytes of the root workspace manifest TOML file and return the path it’s been read from, preferably an absolute path (it will be used as the base path for inherited readmes).