Crate cargo_toml

This crate defines structs that can be deserialized with Serde to load and inspect Cargo.toml metadata.

See Manifest::from_slice.

Structs

Badge

Badges
Deprecated

DependencyDetail

Filesystem

Maintenance

Manifest
The top-level Cargo.toml structure

Package
You can replace Metadata type with your own to parse into something more useful than a generic toml Value

Product
Cargo uses the term “target” for both “target platform” and “build target” (the thing to build), which makes it ambigous. Here Cargo’s bin/lib target is renamed to product.

Profile

Profiles

Target

Workspace

Enums

Dependency

Edition

Error

MaintenanceStatus

Publish

Resolver

Value
Representation of a TOML value.

Traits

AbstractFilesystem

Type Definitions

DepsSet

FeatureSet

PatchSet

TargetDepsSet